c
\name{CompareImpact-package}
\alias{CompareImpact-package}
\docType{package}
\title{
CompareImpact
}
\description{
Simple package for measuring and comparing time-series effect with bayesian prediction modeling
}
\details{
The DESCRIPTION file: CompareImpact.Rmd
approach to integrate multivariate testing and a Bayesian time series prediction model to measure casual inference based on CausalImpact Package 
}
\author{
JeongMin Kwon (cojette@gmail.com)
}
\keyword{ bayesian time series prediction, CompareImpact}